+++ Isolator.Fody.Tests/Tests.cs
@@ -1,0 +1,13 @@
+using Isolator.ConsoleApp;
+
+namespace Isolator.Fody.Tests
+{
+    public class Tests
+    {
+        [Test]
+        public void Program_Main()
+        {
+            Program.Main(Array.Empty<string>());
+        }
+    }
+}